Separate undo logic for tracks; match on scheduled keys instead of indexes
